requirements.txt
fastapi==0.115.12
fitz==0.0.1.dev2
langchain==0.3.25
langchain_groq==0.3.2
python-dotenv==1.1.0
sounddevice==0.5.1
starlette==0.46.2
vosk==0.3.45
